requirements.txt
paho-mqtt
requests
python-dateutil
psycopg[binary, pool]
python-dotenv
black
